Refactor user_config.
Move filter set to user_config without breaking changes.
