+++ index.html
@@ -1598,7 +1598,8 @@
           "report_data": "./js/report_data.js?v=EPOCH",
           "results": "./js/results.js?v=EPOCH",
           "ui_editor": "./js/ui_editor.js?v=EPOCH",
-          "ui_translator": "../js/ui_translator.js?v=EPOCH"
+          "ui_translator": "../js/ui_translator.js?v=EPOCH",
+          "user_config": "../js/user_config.js?v=EPOCH"
         }
       }
     </script>
